+++ core/domain/interface/merchant/shop/shop_rep.go
@@ -10,31 +10,31 @@ package shop
 
 type (
 	IShopRep interface {
-		SaveShop(*Shop) (int, error)
+		SaveShop(*Shop) (int64, error)
 
 		// 商店别名是否存在
-		ShopAliasExists(alias string, shopId int) bool
+		ShopAliasExists(alias string, shopId int64) bool
 
 		// 获取商店值
-		GetValueShop(merchantId, shopId int) *Shop
+		GetValueShop(merchantId, shopId int64) *Shop
 
 		// 获取商户所有商店
-		GetShopsOfMerchant(merchantId int) []Shop
+		GetShopsOfMerchant(merchantId int64) []Shop
 
 		// 删除线上商店
-		DeleteOnlineShop(mchId, shopId int) error
+		DeleteOnlineShop(mchId, shopId int64) error
 
 		// 删除线下门店
-		DeleteOfflineShop(mchId, shopId int) error
+		DeleteOfflineShop(mchId, shopId int64) error
 
 		// 获取线上商店
-		GetOnlineShop(shopId int) *OnlineShop
+		GetOnlineShop(shopId int64) *OnlineShop
 
 		// 保存线上商店
 		SaveOnlineShop(v *OnlineShop, create bool) error
 
 		// 获取线下商店
-		GetOfflineShop(shopId int) *OfflineShop
+		GetOfflineShop(shopId int64) *OfflineShop
 
 		// 保存线下商店
 		SaveOfflineShop(v *OfflineShop, create bool) error
