+++ core/repository/content_rep.go
@@ -29,7 +29,7 @@ func NewContentRep(c db.Connector) content.IContentRep {
 }
 
 // 获取内容
-func (c *contentRep) GetContent(userId int) content.IContent {
+func (c *contentRep) GetContent(userId int64) content.IContent {
 	return contentImpl.NewContent(userId, c)
 }
 
@@ -43,7 +43,7 @@ func (c *contentRep) GetPageById(merchantId, id int) *content.Page {
 }
 
 // 根据标识获取页面
-func (c *contentRep) GetPageByStringIndent(userId int, indent string) *content.Page {
+func (c *contentRep) GetPageByStringIndent(userId int64, indent string) *content.Page {
 	var e content.Page
 	if err := c.Connector.GetOrm().GetBy(&e, "user_id=? and str_indent=?", userId, indent); err == nil {
 		return &e
@@ -52,18 +52,18 @@ func (c *contentRep) GetPageByStringIndent(userId int, indent string) *content.P
 }
 
 // 删除页面
-func (c *contentRep) DeletePage(userId, id int) error {
+func (c *contentRep) DeletePage(userId, id int64) error {
 	_, err := c.Connector.GetOrm().Delete(content.Page{}, "user_id=? AND id=?", userId, id)
 	return err
 }
 
 // 保存页面
-func (c *contentRep) SavePage(userId int, v *content.Page) (int, error) {
+func (c *contentRep) SavePage(userId int64, v *content.Page) (int64, error) {
 	return orm.Save(c.GetOrm(), v, v.Id)
 }
 
 // 获取文章数量
-func (c *contentRep) GetArticleNumByCategory(categoryId int) int {
+func (c *contentRep) GetArticleNumByCategory(categoryId int64) int {
 	num := 0
 	c.Connector.ExecScalar("SELECT COUNT(0) FROM con_article WHERE cat_id=?",
 		&num, categoryId)
@@ -78,7 +78,7 @@ func (c *contentRep) GetAllArticleCategory() []*content.ArticleCategory {
 }
 
 // 判断栏目是否存在
-func (c *contentRep) CategoryExists(indent string, id int) bool {
+func (c *contentRep) CategoryExists(indent string, id int64) bool {
 	num := 0
 	c.Connector.ExecScalar("SELECT COUNT(0) FROM con_category WHERE indent=? and id<>id",
 		&num, indent, id)
@@ -91,12 +91,12 @@ func (c *contentRep) SaveCategory(v *content.ArticleCategory) (id int, err error
 }
 
 // 删除栏目
-func (c *contentRep) DeleteCategory(id int) error {
+func (c *contentRep) DeleteCategory(id int64) error {
 	return c.Connector.GetOrm().DeleteByPk(&content.ArticleCategory{}, id)
 }
 
 // 获取文章
-func (c *contentRep) GetArticleById(id int) *content.Article {
+func (c *contentRep) GetArticleById(id int64) *content.Article {
 	e := content.Article{}
 	if c.Connector.GetOrm().Get(id, &e) == nil {
 		return &e
@@ -105,7 +105,7 @@ func (c *contentRep) GetArticleById(id int) *content.Article {
 }
 
 // 获取文章列表
-func (c *contentRep) GetArticleList(categoryId int, begin int, end int) []*content.Article {
+func (c *contentRep) GetArticleList(categoryId int64, begin int, end int) []*content.Article {
 	list := []*content.Article{}
 	c.Connector.GetOrm().SelectByQuery(&content.Article{},
 		"cat_id=? LIMIT ?,?", categoryId, begin, end-begin)
@@ -118,6 +118,6 @@ func (c *contentRep) SaveArticle(v *content.Article) (i int, err error) {
 }
 
 // 删除文章
-func (c *contentRep) DeleteArticle(id int) error {
+func (c *contentRep) DeleteArticle(id int64) error {
 	return c.Connector.GetOrm().DeleteByPk(&content.Article{}, id)
 }
